Split up angular dashboard directive and controller class (#39188)
* remove unused scope param

* Split up app and app controller into two files

* incorporate docTitle changes, plus type
