+++ src/legacy/core_plugins/kibana/public/dashboard/dashboard_state_manager.ts
@@ -73,11 +73,9 @@ import {
   DashboardAppState,
   SavedDashboardPanel,
   SavedDashboardPanelMap,
-  StagedFilter,
   DashboardAppStateParameters,
+  AddFilterFn,
 } from './types';
-
-export type AddFilterFuntion = ({ field, value, operator, index }: StagedFilter) => void;
 
 /**
  * Dashboard state manager handles connecting angular and redux state between the angular and react portions of the
@@ -100,7 +98,7 @@ export class DashboardStateManager {
   private changeListeners: Array<(status: { dirty: boolean }) => void>;
   private stateMonitor: StateMonitor<DashboardAppStateParameters>;
   private panelIndexPatternMapping: { [key: string]: StaticIndexPattern[] } = {};
-  private addFilter: AddFilterFuntion;
+  private addFilter: AddFilterFn;
   private unsubscribe: () => void;
 
   /**
@@ -119,7 +117,7 @@ export class DashboardStateManager {
     savedDashboard: SavedObjectDashboard;
     AppStateClass: TAppStateClass<DashboardAppState>;
     hideWriteControls: boolean;
-    addFilter: AddFilterFuntion;
+    addFilter: AddFilterFn;
   }) {
     this.savedDashboard = savedDashboard;
     this.hideWriteControls = hideWriteControls;
@@ -321,7 +319,7 @@ export class DashboardStateManager {
 
     const stagedFilters = getStagedFilters(store.getState());
     stagedFilters.forEach(filter => {
-      this.addFilter(filter);
+      this.addFilter(filter, this.getAppState());
     });
     if (stagedFilters.length > 0) {
       this.saveState();
