+++ src/legacy/core_plugins/kibana/public/dashboard/types.ts
@@ -22,6 +22,7 @@ import { AppState } from 'ui/state_management/app_state';
 import { UIRegistry } from 'ui/registry/_registry';
 import { Filter } from '@kbn/es-query';
 import { Query } from 'src/legacy/core_plugins/data/public';
+import { AppState as TAppState } from 'ui/state_management/app_state';
 import { DashboardViewMode } from './dashboard_view_mode';
 
 export interface EmbeddableFactoryRegistry extends UIRegistry<EmbeddableFactory> {
@@ -118,3 +119,30 @@ export interface StagedFilter {
   operator: string;
   index: string;
 }
+
+export type ConfirmModalFn = (
+  message: string,
+  confirmOptions: {
+    onConfirm: () => void;
+    onCancel: () => void;
+    confirmButtonText: string;
+    cancelButtonText: string;
+    defaultFocusedButton: string;
+    title: string;
+  }
+) => void;
+
+export type AddFilterFn = (
+  {
+    field,
+    value,
+    operator,
+    index,
+  }: {
+    field: string;
+    value: string;
+    operator: string;
+    index: string;
+  },
+  appState: TAppState
+) => void;
